requirements.txt
folium==0.15.1
pandas==2.1.4
PyYAML==6.0.1
Requests==2.31.0
